# internal/assets/openapi.yaml
openapi: 3.0.3
info:
  title: File Management API
  description: API for managing files, folders, tags, and semantic search
  version: 1.0.0
  contact:
    name: RxTech Lab
    url: https://github.com/rxtech-lab/files-management
servers:
  - url: /api
    description: API server
...
paths:
...
  /api/files:
    get:
      tags:
        - Files
      summary: List files
      description: Returns a paginated list of files with filtering
      operationId: listFiles
      parameters:
        - name: keyword
          in: query
          description: Search keyword for title, summary, or content
          schema:
            type: string
        - name: folder_id
          in: query
          description: Filter by folder ID
          schema:
            type: integer
        - name: file_type
          in: query
          description: Filter by file type
          schema:
            $ref: '#/components/schemas/FileType'
        - name: tag_ids
          in: query
          description: Filter by tag IDs (comma-separated)
          schema:
            type: string
        - name: status
          in: query
          description: Filter by processing status
          schema:
            $ref: '#/components/schemas/ProcessingStatus'
        - name: sort_by
          in: query
          description: Field to sort by
          schema:
            type: string
            enum: [created_at, updated_at, title, size]
            default: created_at
        - name: sort_order
          in: query
          description: Sort order
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of files
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Files
      summary: Create file
      description: Creates a new file record (after S3 upload)
      operationId: createFile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateFileRequest'
      responses:
        '201':
          description: File created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/files/move:
    post:
      tags:
        - Files
      summary: Move files
      description: Moves multiple files to a target folder
      operationId: moveFiles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveFilesRequest'
      responses:
        '200':
          description: Files moved
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                  - moved_count
                properties:
                  message:
                    type: string
                  moved_count:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/files/batch-download:
    post:
      tags:
        - Files
      summary: Batch download files as ZIP
      description: Downloads multiple files as a ZIP archive (streaming)
      operationId: batchDownloadFiles
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchDownloadRequest'
      responses:
        '200':
          description: ZIP file stream
          content:
            application/zip:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/files/{id}:
    get:
      tags:
        - Files
      summary: Get file
      description: Returns a file by ID with folder and tags
      operationId: getFile
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: File details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Files
      summary: Update file
      description: Updates file metadata
      operationId: updateFile
      parameters:
        - $ref: '#/components/parameters/FileId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateFileRequest'
      responses:
        '200':
          description: File updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Files
      summary: Delete file
      description: Deletes a file and its S3 object
      operationId: deleteFile
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '204':
          description: File deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/files/{id}/process:
    post:
      tags:
        - Files
      summary: Process file
      description: Triggers asynchronous content parsing and embedding generation
      operationId: processFile
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '202':
          description: Processing started
          content:
            application/json:
              schema:
                type: object
                required:
                  - message
                  - status
                properties:
                  message:
                    type: string
                  status:
                    $ref: '#/components/schemas/ProcessingStatus'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/files/{id}/download:
    get:
      tags:
        - Files
      summary: Get file download URL
      description: Returns a presigned download URL for the file
      operationId: getFileDownloadURL
      parameters:
        - $ref: '#/components/parameters/FileId'
      responses:
        '200':
          description: Download URL generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileDownloadResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/files/{id}/tags:
    post:
      tags:
        - Files
      summary: Add tags to file
      description: Adds tags to a file
      operationId: addTagsToFile
      parameters:
        - $ref: '#/components/parameters/FileId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagIdsRequest'
      responses:
        '200':
          description: Tags added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags:
        - Files
      summary: Remove tags from file
      description: Removes tags from a file
      operationId: removeTagsFromFile
      parameters:
        - $ref: '#/components/parameters/FileId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TagIdsRequest'
      responses:
        '200':
          description: Tags removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/File'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # Search
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
  parameters:
...
    FileId:
      name: id
      in: path
      required: true
      description: File ID
      schema:
        type: integer
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 100
    Offset:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
        minimum: 0
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message

    # Tags
    Tag:
      type: object
      required:
        - id
        - user_id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        user_id:
          type: string
        name:
          type: string
        color:
          type: string
          description: Hex color code (e.g., #FF5733)
        description:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    TagIdsRequest:
      type: object
      required:
        - tag_ids
      properties:
        tag_ids:
          type: array
          items:
            type: integer

    # Folders
    Folder:
      type: object
      required:
        - id
        - user_id
        - name
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        user_id:
          type: string
        name:
          type: string
        description:
          type: string
        parent_id:
          type: integer
          nullable: true
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        children:
          type: array
          items:
            $ref: '#/components/schemas/Folder'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    FileType:
      type: string
      enum:
        - music
        - photo
        - video
        - document
        - invoice
    ProcessingStatus:
      type: string
      enum:
        - pending
        - processing
        - completed
        - failed
    File:
      type: object
      required:
        - id
        - user_id
        - title
        - s3_key
        - original_filename
        - file_type
        - processing_status
        - has_embedding
        - created_at
        - updated_at
      properties:
        id:
          type: integer
        user_id:
          type: string
        title:
          type: string
        summary:
          type: string
        content:
          type: string
          description: Parsed text content
        file_type:
          $ref: '#/components/schemas/FileType'
        folder_id:
          type: integer
          nullable: true
        folder:
          $ref: '#/components/schemas/Folder'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        s3_key:
          type: string
        original_filename:
          type: string
        mime_type:
          type: string
        size:
          type: integer
          format: int64
        processing_status:
          $ref: '#/components/schemas/ProcessingStatus'
        processing_error:
          type: string
        has_embedding:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateFileRequest:
      type: object
      required:
        - title
        - s3_key
        - original_filename
      properties:
        title:
          type: string
        s3_key:
          type: string
        original_filename:
          type: string
        mime_type:
          type: string
        size:
          type: integer
          format: int64
        folder_id:
          type: integer
          nullable: true
        file_type:
          $ref: '#/components/schemas/FileType'
    UpdateFileRequest:
      type: object
      properties:
        title:
          type: string
        summary:
          type: string
        file_type:
          $ref: '#/components/schemas/FileType'
        folder_id:
          type: integer
          nullable: true
    MoveFilesRequest:
      type: object
      required:
        - file_ids
      properties:
        file_ids:
          type: array
          items:
            type: integer
        folder_id:
          type: integer
          nullable: true
          description: Target folder ID (null for root)
    BatchDownloadRequest:
      type: object
      required:
        - file_ids
      properties:
        file_ids:
          type: array
          items:
            type: integer
          description: Array of file IDs to download
    FileListResponse:
      type: object
      required:
        - data
        - total
        - limit
        - offset
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/File'
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    FileDownloadResponse:
      type: object
      required:
        - download_url
        - key
        - filename
        - expires_at
      properties:
        download_url:
          type: string
          format: uri
        key:
          type: string
        filename:
          type: string
        expires_at:
          type: string
          format: date-time

    # Search
...
security:
  - BearerAuth: []